Add `TestrailProject#get_plans` ability to filter
